+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.NontonKuyMade"
         android:usesCleartextTraffic="true">
-        <activity android:name=".core.ui.detail.DetailMovieActivity"></activity>
+        <activity android:name=".ui.detail.DetailMovieActivity"></activity>
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
