+++ app/src/main/java/com/example/nontonkuymade/core/ui/MovieAdapter.kt
@@ -1,14 +1,12 @@
 package com.example.nontonkuymade.core.ui
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.core.databinding.ItemGridBinding
 import com.example.nontonkuymade.BuildConfig
-import com.example.nontonkuymade.R
 import com.example.nontonkuymade.core.data.source.local.entity.MovieEntity
-import com.example.nontonkuymade.databinding.ItemGridBinding
 
 class MovieAdapter : RecyclerView.Adapter<MovieAdapter.ViewHolder>(){
