+++ app/src/main/java/com/example/nontonkuymade/ui/favorite/MovieFavoriteViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.nontonkuymade.core.ui.favorite
+package com.example.nontonkuymade.ui.favorite
 
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.ViewModel
