+++ app/src/main/res/layout/fragment_favorite.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".core.ui.favorite.FavoriteFragment">
+    tools:context=".ui.favorite.FavoriteFragment">
 
     <ProgressBar
         android:id="@+id/pb_fav_mov"
