+++ core/src/main/res/layout/item_grid.xml
