+++ core/src/main/res/layout/item_horizontal.xml
